+++ packages/model-typings/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/model-typings
 
+## 0.3.8
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.7.1
+  </details>
+
 ## 0.3.7
 
 ### Patch Changes